# ProyectoNotas/src/openapi/Swagger_ProyectoNotas_API.yaml
openapi: 3.0.0
info:
  title: Proyecto Notas API
  description: API para gestionar notas, incluyendo crear, listar, actualizar, eliminar y subir archivos.
  version: 1.0.0

servers:
  - url: http://localhost:3100
    description: Servidor local

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

paths:
  /api/notes:
    get:
      summary: Listar todas las notas
      description: Obtener todas las notas con filtros, ordenación y paginación.
      security:
        - bearerAuth: []
      parameters:
        - name: filter[title]
          in: query
          description: Filtrar notas por título.
          schema:
            type: string
        - name: filter[content]
          in: query
          description: Filtrar notas por contenido.
          schema:
            type: string
        - name: filter[dateRange][startDate]
          in: query
          description: Fecha de inicio del rango de creación.
          schema:
            type: string
            format: date-time
        - name: filter[dateRange][endDate]
          in: query
          description: Fecha de fin del rango de creación.
          schema:
            type: string
            format: date-time
        - name: sortBy
          in: query
          description: Ordenar por campo (`createdAt`, `updatedAt`, `title`, `size`).
          schema:
            type: string
            enum: [createdAt, updatedAt, title, size]
        - name: order
          in: query
          description: Orden ascendente (`asc`) o descendente (`desc`).
          schema:
            type: string
            enum: [asc, desc]
        - name: page
          in: query
          description: Número de página para paginación.
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Cantidad de notas por página.
          schema:
            type: integer
            default: 5
      responses:
        '200':
          description: Lista de notas obtenida con éxito.
        '400':
          description: Error en los parámetros de entrada.
        '401':
          description: No autorizado.

    post:
      summary: Crear una nota
      description: Crear una nueva nota con nombre y contenido.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: DavidGuisado
                content:
                  type: string
                  example: Soy David Guisado Núñez.
      responses:
        '201':
          description: Nota creada exitosamente.
        '400':
          description: Error en los datos de entrada.
        '401':
          description: No autorizado.

  /api/notes/{name}:
    put:
      summary: Actualizar una nota
      description: Actualizar el contenido de una nota existente por nombre.
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: path
          required: true
          description: Nombre de la nota a actualizar.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                  example: Este es el contenido actualizado de la nota.
      responses:
        '200':
          description: Nota actualizada exitosamente.
        '404':
          description: Nota no encontrada.
        '401':
          description: No autorizado.

    delete:
      summary: Eliminar una nota
      description: Eliminar una nota por nombre.
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: path
          required: true
          description: Nombre de la nota a eliminar.
          schema:
            type: string
      responses:
        '200':
          description: Nota eliminada exitosamente.
        '404':
          description: Nota no encontrada.
        '401':
          description: No autorizado.

  /api/notes/upload:
    post:
      summary: Subir notas
      description: Subir archivos `.note` al servidor.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '200':
          description: Archivos subidos exitosamente.
        '400':
          description: Error en la subida de archivos.
        '401':
          description: No autorizado.

  /api/notes/download/{filename}:
    get:
      summary: Descargar una nota
      description: Descargar un archivo `.note` específico.
      parameters:
        - name: filename
          in: path
          required: true
          description: Nombre del archivo a descargar.
          schema:
            type: string
      responses:
        '200':
          description: Archivo descargado exitosamente.
        '404':
          description: Archivo no encontrado.